Removed unnecessary globals from main
