+++ src/main.cpp
@@ -47,9 +47,9 @@
 #include <stb_image.h>
 
 // Headers locais, definidos na pasta "include/"
-#include "utils.h"
+
 #include "matrices.h"
-#include "textrendering.h"
+
 
 #include "ObjModel.h"
 #include "Camera.h"
@@ -77,9 +77,6 @@ GLint g_bbox_max_uniform;
 void PushMatrix(glm::mat4 M);
 void PopMatrix(glm::mat4 &M);
 
-// Funções abaixo renderizam como texto na janela OpenGL algumas matrizes e
-// outras informações do programa. Definidas após main().
-void TextRendering_ShowFramesPerSecond(GLFWwindow *window);
 
 // Funções callback para comunicação com o sistema operacional e interação do
 // usuário. Veja mais comentários nas definições das mesmas, abaixo.
@@ -96,12 +93,6 @@ std::stack<glm::mat4> g_MatrixStack;
 // Razão de proporção da janela (largura/altura). Veja função FramebufferSizeCallback().
 float g_ScreenRatio = 1.0f;
 
-// "g_LeftMouseButtonPressed = true" se o usuário está com o botão esquerdo do mouse
-// pressionado no momento atual. Veja função MouseButtonCallback().
-bool g_LeftMouseButtonPressed = false;
-bool g_RightMouseButtonPressed = false;  // Análogo para botão direito do mouse
-bool g_MiddleMouseButtonPressed = false; // Análogo para botão do meio do mouse
-
 // LookAtCamera camera(0.0f, 0.0f, 3.5f);
 FreeCamera camera(3.14f, -0.7f, glm::vec4(1.7f, 0.2f, 1.7f, 1.0f), 5.0f);
 texture::TextureLoader textureLoader;
@@ -114,9 +105,6 @@ GLuint LoadShader_Fragment(const char *filename);                            // 
 void LoadShader(const char *filename, GLuint shader_id);                     // Função utilizada pelas duas acima
 GLuint CreateGpuProgram(GLuint vertex_shader_id, GLuint fragment_shader_id); // Cria um programa de GPU
 
-
-// Variável que controla se o texto informativo será mostrado na tela.
-bool g_ShowInfoText = true;
 
 int main(int argc, char *argv[])
 {
@@ -174,48 +162,6 @@ void PopMatrix(glm::mat4 &M)
         M = g_MatrixStack.top();
         g_MatrixStack.pop();
     }
-}
-
-
-// Variáveis globais que armazenam a última posição do cursor do mouse, para
-// que possamos calcular quanto que o mouse se movimentou entre dois instantes
-// de tempo. Utilizadas no callback CursorPosCallback() abaixo.
-double g_LastCursorPosX, g_LastCursorPosY;
-
-// Escrevemos na tela o número de quadros renderizados por segundo (frames per
-// second).
-void TextRendering_ShowFramesPerSecond(GLFWwindow *window)
-{
-    if (!g_ShowInfoText)
-        return;
-
-    // Variáveis estáticas (static) mantém seus valores entre chamadas
-    // subsequentes da função!
-    static float old_seconds = (float)glfwGetTime();
-    static int ellapsed_frames = 0;
-    static char buffer[20] = "?? fps";
-    static int numchars = 7;
-
-    ellapsed_frames += 1;
-
-    // Recuperamos o número de segundos que passou desde a execução do programa
-    float seconds = (float)glfwGetTime();
-
-    // Número de segundos desde o último cálculo do fps
-    float ellapsed_seconds = seconds - old_seconds;
-
-    if (ellapsed_seconds > 1.0f)
-    {
-        numchars = snprintf(buffer, 20, "%.2f fps", ellapsed_frames / ellapsed_seconds);
-
-        old_seconds = seconds;
-        ellapsed_frames = 0;
-    }
-
-    float lineheight = TextRendering_LineHeight(window);
-    float charwidth = TextRendering_CharWidth(window);
-
-    TextRendering_PrintString(window, buffer, 1.0f - (numchars + 1) * charwidth, 1.0f - lineheight, 1.0f);
 }
 
 // Função que carrega os shaders de vértices e de fragmentos que serão
